pyproject.toml
[tool.poetry]
name = "ozon-demo-project"
version = "0.1.0"
description = ""
authors = ["Anhelina Babareka <anhielina.babareka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.2"
selene = "2.0.0rc4"
allure-pytest = "^2.13.2"
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
